* fix, MongoDB plugin: improved validation of mongo_table value, expecting   it in <database>.<collection> format. Updated docs. * fix, MongoDB plugin: MongoDB_create_indexes() does not resolve dynamic   table again but relies on centralized name composition. * fix, print plugin: if files_uid and/or files_gid are in use, pointer to   latest file now does a lchown() to the configured username and/or group. * fix, print plugin: plugin cache can now grow beyond 2GB.

+++ src/print_plugin.h
@@ -31,8 +31,8 @@
 struct scratch_area {
   unsigned char *base;
   unsigned char *ptr;
-  int num;
-  int size;
+  u_int64_t num;
+  u_int64_t size;
   struct scratch_area *next;
 };
